Fix #1 - allow for GIF files larger than 109x109px
While there were WIDTH and HEIGHT constants in the code, a lot of places in SVG.resx either had 109px hardcoded or assumed that width and height were always going to be 109px.

Additionally, KanjiVG strokes are hardcoded to 109px but luckily SVG has a scale transformation that makes this a non-issue.

Fixes #1